XIVY-11599 Intermediate events
