+++ engine-cockpit/webContent/includes/components/monitor/events/intermediate/bean.xhtml
@@ -5,44 +5,42 @@
   xmlns:cc="http://java.sun.com/jsf/composite/composite">
 
 <div class="card">
-  <h:form id="startEventBean">
+  <h:form id="startEventFringForm">
     <div class="card-header">
       <div class="card-title p-d-flex p-ai-center">
         <i class="p-pr-3 si si-cog"></i>
         <h6 class="p-m-0">Bean</h6>
         <p:commandButton id="start" title="Starts the bean" icon="si si-controls-play"
-            action="#{startEventDetailBean.start()}" update="@form" styleClass="p-col-fixed p-ml-auto ui-button-success rounded-button" disabled="#{startEventDetailBean.running}"/>
+            action="#{intermediateEventDetailBean.event.start()}" update="@form"
+            styleClass="p-col-fixed p-ml-auto ui-button-success rounded-button" disabled="#{intermediateEventDetailBean.event.running}"/>
         <p:commandButton id="stop" title="Stops the bean" icon="si si-controls-stop"
-            action="#{startEventDetailBean.stop()}" update="@form" styleClass="p-col-fixed p-ml-1 ui-button-danger rounded-button" disabled="#{!startEventDetailBean.running}"/>
+            action="#{intermediateEventDetailBean.event.stop()}" update="@form"
+            styleClass="p-col-fixed p-ml-1 ui-button-danger rounded-button" disabled="#{!intermediateEventDetailBean.event.running}"/>
       </div>
     </div>
     
     <p:panelGrid columns="2" columnClasses="p-col-12 p-md-3 p-lg-3, p-col-12 p-md-9 p-lg-9" layout="flex"
       styleClass="grey-label-panel ui-fluid">
-      <p:outputLabel for="class" value="Class" />
-      <h:outputText id="class" value="#{startEventDetailBean.beanClass}"/>
       <p:outputLabel for="name" value="Name" />
-      <h:outputText id="name" value="#{startEventDetailBean.beanName}"/>
+      <h:outputText id="name" value="#{intermediateEventDetailBean.event.beanName}"/>
       <p:outputLabel for="description" value="Description" />
-      <h:outputText id="description" value="#{startEventDetailBean.beanDescription}"/>
+      <h:outputText id="description" value="#{intermediateEventDetailBean.event.beanDescription}"/>
       <p:outputLabel for="requestPath" value="Request Path" />
-      <h:outputText id="requestPath" value="#{startEventDetailBean.fullRequestPath}"/>
+      <h:outputText id="requestPath" value="#{intermediateEventDetailBean.event.fullRequestPath}"/>
       <p:outputLabel for="serviceState" value="State" />
-      <h:outputText id="serviceState" value="#{startEventDetailBean.serviceState}"/>
-      <p:outputLabel for="configuration" value="Configuration" />
-      <h:outputText id="configuration" value="#{startEventDetailBean.beanConfiguration}"/>
+      <h:outputText id="serviceState" value="#{intermediateEventDetailBean.event.serviceState}"/>
 
       <p:outputLabel for="lastStart" value="Last Start" />
-      <h:outputText id="lastStart" value="#{startEventDetailBean.lastStartTimestamp}"/>
+      <h:outputText id="lastStart" value="#{intermediateEventDetailBean.event.lastStartTimestamp}"/>
 
       <p:outputLabel for="lastStartError" value="Last Start Error" />
-      <cc:Error id="lastStartError" error="#{startEventDetailBean.lastStartError}"/>
+      <cc:Error id="lastStartError" error="#{intermediateEventDetailBean.event.lastStartError}"/>
 
       <p:outputLabel for="lastInitError" value="Last Initialization Error" />
-      <cc:Error id="lastInitError" error="#{startEventDetailBean.lastInitializationError}"/>
+      <cc:Error id="lastInitError" error="#{intermediateEventDetailBean.event.lastInitializationError}"/>
 
       <p:outputLabel for="lastStopError" value="Last Stop Error" />
-      <cc:Error id="lastStopError" error="#{startEventDetailBean.lastStopError}"/>
+      <cc:Error id="lastStopError" error="#{intermediateEventDetailBean.event.lastStopError}"/>
     </p:panelGrid>
   </h:form>
 </div>
