--- engine-cockpit/webContent/includes/components/monitor/startevent-firings.xhtml
@@ -1,46 +1,0 @@
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:f="http://xmlns.jcp.org/jsf/core"
-  xmlns:h="http://xmlns.jcp.org/jsf/html" xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-  xmlns:c="http://xmlns.jcp.org/jsp/jstl/core" xmlns:ic="http://ivyteam.ch/jsf/component"
-  xmlns:p="http://primefaces.org/ui" xmlns:pe="http://primefaces.org/ui/extensions"
-  xmlns:cc="http://java.sun.com/jsf/composite/composite">
-
-<div class="card">
-  <h:form id="startEventFiring">
-    <div class="card-header">
-      <div class="card-title p-d-flex p-ai-center">
-        <i class="p-pr-3 si si-navigation-right-circle-1"></i>
-        <h6 class="p-m-0">Executions</h6>
-      </div>
-    </div>
-    
-    <p:panelGrid columns="6" columnClasses="p-col-12 p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3, p-col-12- p-md-1 p-lg-1, p-col-12- p-md-1 p-lg-3, p-col-12- p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3" layout="flex"
-      styleClass="grey-label-panel ui-fluid">
-      <p:outputLabel for="executions" value="Executions" />
-      <h:outputText id="executions" value="#{startEventDetailBean.executions}"/>
-      <p:outputLabel for="duration" value="Execution Duration" />
-      <h:panelGroup>
-        <h:outputText id="duration" value="#{startEventDetailBean.minExecutionTime} / #{startEventDetailBean.avgExecutionTime} / #{startEventDetailBean.maxExecutionTime}"/>
-        <p:tooltip for="duration" value="min / avg / max"/>
-      </h:panelGroup>
-      <p:outputLabel for="errors" value="Errors" />
-      <h:outputText id="errors" value="#{startEventDetailBean.errors}"/>
-    </p:panelGrid>
-    <p:dataTable value="#{startEventDetailBean.firings}" var="firing" widgetVar="firingTable" styleClass="ui-fluid"
-            paginator="true" rows="20" paginatorPosition="bottom" paginatorAlwaysVisible="false"
-            id="firingTable" style="min-width: 950px;">
-      <p:column headerText="Timestamp" sortBy="#{firing.timestamp}" filterBy="#{firing.timestamp}">
-        <h:outputText value="#{firing.timestamp}"/>
-      </p:column>
-      <p:column headerText="Duration" sortBy="#{firing.duration}" filterBy="#{firing.duration}">
-        <h:outputText value="#{firing.duration}"/>
-      </p:column>
-      <p:column headerText="Reason" sortBy="#{firing.reason}" filterBy="#{firing.reason}">
-        <h:outputText value="#{firing.reason}"/>
-      </p:column>
-      <p:column headerText="Error" sortBy="#{firing.error}" filterBy="#{firing.error}">
-        <cc:Error id="error" error="#{firing.error}"/>
-      </p:column>
-    </p:dataTable>
-  </h:form>
-</div>
-</html>
