+++ engine-cockpit/webContent/resources/composite/Firings.xhtml
@@ -1,0 +1,55 @@
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:f="http://xmlns.jcp.org/jsf/core"
+  xmlns:h="http://xmlns.jcp.org/jsf/html" xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+  xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
+  xmlns:p="http://primefaces.org/ui">
+<cc:interface>
+  <cc:attribute name="event" />
+</cc:interface>
+
+<cc:implementation>
+  <div class="card">
+    <h:form id="eventFiring">
+      <div class="card-header">
+        <div class="card-title p-d-flex p-ai-center">
+          <i class="p-pr-3 si si-navigation-right-circle-1"></i>
+          <h6 class="p-m-0">Executions</h6>
+        </div>
+      </div>
+
+      <p:panelGrid columns="6"
+        columnClasses="p-col-12 p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3, p-col-12- p-md-1 p-lg-1, p-col-12- p-md-1 p-lg-3, p-col-12- p-md-12 p-lg-1, p-col-12- p-md-12 p-lg-3"
+        layout="flex" styleClass="grey-label-panel ui-fluid">
+        <p:outputLabel for="executions" value="Executions" />
+        <h:outputText id="executions" value="#{cc.attrs.event.executions}" />
+        <p:outputLabel for="duration" value="Execution Duration" />
+        <h:panelGroup>
+          <h:outputText id="duration"
+            value="#{cc.attrs.event.minExecutionTime} / #{cc.attrs.event.avgExecutionTime} / #{cc.attrs.event.maxExecutionTime}" />
+          <p:tooltip for="duration" value="min / avg / max"/>
+        </h:panelGroup>
+        <p:outputLabel for="errors" value="Errors" />
+        <h:outputText id="errors" value="#{cc.attrs.event.errors}" />
+      </p:panelGrid>
+      <p:dataTable value="#{cc.attrs.event.firings}" var="firing" widgetVar="firingTable" styleClass="ui-fluid" paginator="true"
+        rows="20" paginatorPosition="bottom" paginatorAlwaysVisible="false" id="firingTable" style="min-width: 950px;">
+        <p:column headerText="Timestamp" sortBy="#{firing.timestamp}" filterBy="#{firing.timestamp}">
+          <h:outputText value="#{firing.timestamp}" />
+        </p:column>
+        <p:column headerText="Duration" sortBy="#{firing.duration}" filterBy="#{firing.duration}">
+          <h:outputText value="#{firing.duration}" />
+        </p:column>
+        <p:column headerText="Reason" sortBy="#{firing.reason}" filterBy="#{firing.reason}">
+          <h:outputText value="#{firing.reason}" />
+        </p:column>
+        <p:column headerText="Error" sortBy="#{firing.error}" filterBy="#{firing.error}">
+          <h:panelGroup styleClass="p-d-flex">
+            <h:outputText id="error" value="#{firing.error.message}" style="width: 100%;"/>
+            <p:commandButton value="..." rendered="#{firing.error.available}" styleClass="p-ml-1 rounded-button" title="Show full error" 
+                update="errorDialog" action="#{errorBean.setSelected(cc.attrs.error)}" oncomplete="PF('errorDialog').show();"/>
+          </h:panelGroup>
+        </p:column>
+      </p:dataTable>
+    </h:form>
+  </div>
+</cc:implementation>
+</html>
